Move file `__enter__`/`__exit__` to the Python layer (#2087)
* Move file __enter__/__exit__ to Python level

* Demo update

* Format updates

* Change RuntimeError to ImportError

* Tidy up imports

* flake8 fix

* Try #noqa again
