+++ python/dolfinx/io.py
@@ -13,15 +13,14 @@ import numpy as np
 
 import ufl
 from dolfinx import cpp as _cpp
-from dolfinx.cpp.io import distribute_entity_data
-from dolfinx.cpp.io import perm_gmsh as cell_perm_gmsh
+from dolfinx.cpp.io import distribute_entity_data  # noqa: F401
+from dolfinx.cpp.io import perm_gmsh as cell_perm_gmsh  # noqa F401
 from dolfinx.fem import Function
 from dolfinx.mesh import GhostMode, Mesh
 
 from mpi4py import MPI as _MPI
 
-__all__ = ["FidesWriter", "VTKFile", "VTXWriter", "XDMFFile", "cell_perm_gmsh",
-           "distribute_entity_data"]
+__all__ = ["FidesWriter", "VTKFile", "VTXWriter", "XDMFFile", "cell_perm_gmsh", "distribute_entity_data"]
 
 
 def _extract_cpp_functions(functions: typing.Union[typing.List[Function], Function]):
@@ -32,7 +31,11 @@ def _extract_cpp_functions(functions: typing.Union[typing.List[Function], Functi
         return [getattr(functions, "_cpp_object", functions)]
 
 
-try:
+if _cpp.common.has_adios2:
+    # FidesWriter and VTXWriter require ADIOS2
+
+    __all__ = __all__ + ["FidesWriter", "VTXWriter"]
+
     class VTXWriter(_cpp.io.VTXWriter):
         """Interface to VTK files for ADIOS2
 
@@ -67,6 +70,12 @@ try:
                 # Input is a single function or a list of functions
                 super().__init__(comm, filename, _extract_cpp_functions(output))
 
+        def __enter__(self):
+            return self
+
+        def __exit__(self, exception_type, exception_value, traceback):
+            self.close()
+
     class FidesWriter(_cpp.io.FidesWriter):
         """Interface to Fides file formt.
 
@@ -97,14 +106,12 @@ try:
                 super().__init__(comm, filename, output)
             except (NotImplementedError, TypeError):
                 super().__init__(comm, filename, _extract_cpp_functions(output))
-except AttributeError:
-    class FidesWriter():
-        def __init__(self, *args):
-            raise RuntimeError("DOLFINx has not been configured with ADIOS2 support")
 
-    class VTXWriter():
-        def __init__(self, *args):
-            raise RuntimeError("DOLFINx has not been configured with ADIOS2 support")
+        def __enter__(self):
+            return self
+
+        def __exit__(self, exception_type, exception_value, traceback):
+            self.close()
 
 
 class VTKFile(_cpp.io.VTKFile):
@@ -116,6 +123,12 @@ class VTKFile(_cpp.io.VTKFile):
 
     """
 
+    def __enter__(self):
+        return self
+
+    def __exit__(self, exception_type, exception_value, traceback):
+        self.close()
+
     def write_mesh(self, mesh: Mesh, t: float = 0.0) -> None:
         """Write mesh to file for a given time (default 0.0)"""
         self.write(mesh, t)
@@ -126,6 +139,12 @@ class VTKFile(_cpp.io.VTKFile):
 
 
 class XDMFFile(_cpp.io.XDMFFile):
+    def __enter__(self):
+        return self
+
+    def __exit__(self, exception_type, exception_value, traceback):
+        self.close()
+
     def write_mesh(self, mesh: Mesh) -> None:
         """Write mesh to file for a given time (default 0.0)"""
         super().write_mesh(mesh)
